Mutualize monitoring autoconfig
