+++ internal/monitoring/monitoring.go
@@ -1,0 +1,17 @@
+package monitoring
+
+import (
+	"os"
+
+	"github.com/sirupsen/logrus"
+)
+
+func init() {
+	if logLevelStr := os.Getenv("LOG_LEVEL"); logLevelStr != "" {
+		logLevel, err := logrus.ParseLevel(logLevelStr)
+		if err != nil {
+			logrus.Fatal(err)
+		}
+		logrus.SetLevel(logLevel)
+	}
+}
